==> documents/mt4/NSRI-main 3/2. Data inspection.Rmd <==
---
title: "2. Data inspection"
author: "Gerko Vink"
date: ""
output: 
   html_document:
    toc: true
    toc_depth: 5
    toc_float: true
    number_sections: false
---
  
<style type="text/css">
  
body{ /* Normal  */
  font-size: 12px;
  }
td {  /* Table  */
  font-size: 12px;
}
h1.title {
  font-size: 18px;
  color: DarkBlue;
}
h1 { /* Header 1 */
  font-size: 18px;
}
h2 { /* Header 2 */
  font-size: 18px;
}
h3 { /* Header 3 */
  font-size: 18px;
}
code.r{ /* Code block */
  font-size: 12px;
}
pre { /* Code block - determines code spacing between lines */
  font-size: 14px;
}
</style>

---

This document is aimed at inspecting the data and detecting anomalies that may influence any data imputation and subsequent data analysis steps. The aim for this document is not to provide a fix if any peculiarities and/or anomalies are found. 

---

# Packages used
For this document we use the following packages
```{r message=FALSE, warning=FALSE}
library(dplyr)    # Data manipulation
library(mice)     # Data imputation
library(magrittr) # Pipes
library(purrr)    # Functional programming
library(haven)    # Import data
library(DT)       # Interactive tables
```

---

# Read in the data file
We read in the workspace `1. Data import and preparation.Rdata` from the `\Workspaces\` directory in the project's root. This workspace has been generated by the document `1. Data import and preparation.Rmd`. 
```{r}
load("Workspaces/1. Data import and preparation.Rdata")
```


---

# Dimensions of the data
To obtain the dimensionality of the data:
```{r}
dim(NSRI_df)
```
We can see that the object `NSRI_df` has `r dim(NSRI_df)[1]` rows and `r dim(NSRI_df)[2]` columns. 

---

# Are the row ID's unique?
```{r}
NSRI_df %$% unique(ID) %>% length 
NSRI_df %$% length(ID)
```
The length of the vector of unique `ID`'s is equal to the length of the vector `ID`. There are no duplicate `ID`'s: every `ID` is unique. 

---

# Are there any duplicate rows
```{r}
(duplicated(NSRI_df) | duplicated(NSRI_df, fromLast = TRUE)) %>% 
  any()
```
There are no duplicate rows in the data. 

---

# Measurement level of the columns
```{r}
NSRI_df %>% map(class) %>% unlist
```

---

# Distribution of the cases over `V`
To inspect how the cases are distributed over column `V` - the indicator for the 
planned missingness - we use
```{r}
NSRI_df %$%
  table(V)
```
We can see that the randomization resulted in approximately equal group sizes. This may indicate that randomization was succesful. To inspect the proportional distribution of the cases over `V`:
```{r}
NSRI_df %$%
  table(V) %>% 
  prop.table %>% 
  round(digits = 3)
```
The versions `V` over the cases are approximately distributed in thirds. 

---

# Missingness inspection

---

## Number of missing values
To inspect the number of missing values in the data we first calculate the total number of cells:
```{r}
ncells <- nrow(NSRI_df) * ncol(NSRI_df)
ncells
```
Out of `r nrow(NSRI_df) * ncol(NSRI_df)` cells, a total of 
```{r}
nmis <- is.na(NSRI_df) %>% sum
nmis
```
are missing, which translates to 
```{r}
nmis / ncells
```
`r round(nmis / ncells, 3)*100` percent of cells being unobserved in the data. 

---

## Missings per column
```{r}
apply(NSRI_df, 2, function(x) sum(is.na(x)))
```

---

## Completely unobserved columns
```{r}
which <- apply(NSRI_df, 2, function(x) all(is.na(x)))
names(NSRI_df)[which]
```
Column `IC2` is completely unobserved. Missingness is only observed on items where missingness is expected due to the questionnaire design. 

---

## Missing data pattern
```{r, fig.width = 30}
md.pattern(NSRI_df[, 11:85], rotate.names = TRUE)
```

The missing data pattern displays 3 distinct patterns; each with 50 columns observed and 25 columns unobserved. This is conform the expected planned missingness. 

---

# Demographic comparisons

---

## Version 
To quickly identify if there are differences in the medians between the randomly assigned versions `V`. 
```{r}
NSRI_df %>% 
  group_by(V) %>% 
  summarise_if(is.numeric, median, na.rm = TRUE) %>% 
  datatable(options = list(scrollX = TRUE))
```

---

## Research involvement
To quickly identify if there are differences in the medians for groups that do or do not spend more than 8 hours on research a week `D1`. 
```{r}
NSRI_df %$% table(D1)
```
Column `D1` is constant. 

---

## Academic field
To quickly identify if there are differences in the medians between the academic fields `D2`. 
```{r}
NSRI_df %$% table(D2)
NSRI_df %>% 
  na_if(8) %>% # Exclude the not applicable's
  group_by(D2) %>% 
  summarise_if(is.numeric, median, na.rm = TRUE) %>% 
  datatable(options = list(scrollX = TRUE))
```

--- 

## Academic rank 
To quickly identify if there are differences in the medians between the rank `D3`. 
```{r}
NSRI_df %$% table(D3)
```
There are no observations for `none of the above`. Let's refactor the corresponding column to avoid redundant comparisons later on.
```{r}
NSRI_df %<>% mutate(D3 = factor(D3)) 
```

```{r}
NSRI_df %>% 
  na_if(8) %>% # Exclude the not applicable's
  group_by(D3) %>% 
  summarise_if(is.numeric, median, na.rm = TRUE) %>% 
  datatable(options = list(scrollX = TRUE))
```

---

## Gender
To quickly identify if there are differences in the medians between the rank `D3`. 
```{r}
NSRI_df %$% table(D4)
NSRI_df %>% 
  na_if(8) %>% # Exclude the not applicable's
  group_by(D4) %>% 
  summarise_if(is.numeric, median, na.rm = TRUE) %>% 
  datatable(options = list(scrollX = TRUE))
```

---

## Engagement
To quickly identify if there are differences in the medians between the rank `D3`. 
```{r}
NSRI_df %$% table(D5)
NSRI_df %>% 
  na_if(8) %>% # Exclude the not applicable's
  group_by(D5) %>% 
  summarise_if(is.numeric, median, na.rm = TRUE) %>% 
  datatable(options = list(scrollX = TRUE))
```

---

# Demographic distribution over V

---

## V x Field D2
```{r}
NSRI_df %$% 
  table(V, D2) %>% 
  datatable
```

---

## V x Rank D3
```{r}
NSRI_df %$% 
  table(V, D3) %>% 
  datatable
```

---

## V x Gender D4
```{r}
NSRI_df %$% 
  table(V, D4) %>% 
  datatable
```

---

## V x Engagement D5
```{r}
NSRI_df %$% 
  table(V, D5) %>% 
  datatable
```

---

# Demographic distribution over Field

---

## Field D2 x Rank D3
```{r}
NSRI_df %$% 
  table(D2, D3) %>% 
  datatable
```

---

## Field D2 x Gender D4
```{r}
NSRI_df %$% 
  table(D2, D4) %>% 
  datatable
```

---

## Field D2 x Engagement D5
```{r}
NSRI_df %$% 
  table(D2, D5) %>% 
  datatable
```

---

# Demographic distribution over Rank

---

## Rank D3 x Gender D4
```{r}
NSRI_df %$% 
  table(D2, D4) %>% 
  datatable
```

---

## Rank D3 x Engagement D5
```{r}
NSRI_df %$% 
  table(D2, D5) %>% 
  datatable
```

---

# Demographic distribution over Gender


---

## Gender x Engagement D5
```{r}
NSRI_df %$% 
  table(D2, D5) %>% 
  datatable
```

---

# Effective sample size

---

## Field D2 x Rank D3
```{r}
NSRI_df %>% 
  group_by(D2, D3) %>% 
  tally %>% 
  datatable
```

---

## Field D2 x Rank D3 x gender D4
```{r}
NSRI_df %>% 
  group_by(D2, D3, D4) %>% 
  tally %>% 
  datatable
```

---

# Not Applicable

---

## Recode the not applicable's
The `not applicables` have been coded as 9 in the `SPSS` set. In the data frame `NSRI_df` the not applicables have been refactored to value `8`. Conceptually, for some - but not all - relevant items, the value `0` would reflect the nature of not applicable better. Therefore, we already recode the `8`'s to `0`'s. 
```{r}
NSRI_df[NSRI_df == 8] <- 0
```

---

## How many not applicables per column
```{r}
(NSRI_df == 0) %>% colSums(na.rm = TRUE)
```

---

# Exporting the workspace
We export the objects `NSRI_df` and `raw_data` in an `R` workspace to the `\Workspaces\` directory in the project root. The resulting workspace name mirrors the name of this document. 
```{r}
save(NSRI_df, raw_data, file = "Workspaces/2. Data inspection.Rdata")
```

---

End of document. 

---

```{r}
sessionInfo()
```
