==> documents/mt4/NSRI-main 3/3. Data Imputation.Rmd <==
---
title: "3. Data imputation"
author: "Gerko Vink"
date: ""
bibliography: references.bib
output: 
   html_document:
    toc: true
    toc_depth: 5
    toc_float: true
    number_sections: false
---
  
<style type="text/css">
  
body{ /* Normal  */
  font-size: 12px;
  }
td {  /* Table  */
  font-size: 12px;
}
h1.title {
  font-size: 18px;
  color: DarkBlue;
}
h1 { /* Header 1 */
  font-size: 18px;
}
h2 { /* Header 2 */
  font-size: 18px;
}
h3 { /* Header 3 */
  font-size: 18px;
}
code.r{ /* Code block */
  font-size: 12px;
}
pre { /* Code block - determines code spacing between lines */
  font-size: 14px;
}
</style>

---

This document is aimed at imputing the planned missingness in the data. 

---

# Packages used
For this document we use the following packages
```{r message=FALSE, warning=FALSE}
library(dplyr)    # Data manipulation
library(mice)     # Data imputation
library(magrittr) # Pipes
library(purrr)    # Functional programming
library(haven)    # Import data
library(DT)       # Interactive tables
```

---

# Read in the data file
We read in the workspace `2. Data inspection.Rdata` from the `\Workspaces\` directory in the project's root. This workspace has been generated by the document `2. Data inspection.Rmd`. The proces leading to this workspace is also documented in the file `2.-Data-inspection.html`.
```{r}
load("Workspaces/2. Data inspection.Rdata")
```

---

# Setting the RNG seed
Because we use the random number generator during imputation, we fix the RNG seed. 
```{r}
set.seed(123)
```
Fixing the seed allows for full reproducibility in the imputation procedure. 

---

# Imputation

We use package `mice` [@micevbuuren] in `R` [@R-base]

---

## Make predictor matrix
Not every column holds information that can - or should - be used for imputation. For example, columns `D1`, `IC1` and `IC2` are constant and column `ID` is redundant for the process of imputation. Further, the test items for the randomized response are also irrelevant for imputation. The predictor matrix in mice serves as the basis for the modeling of the incomplete columns. 

First, we create a default predictor matrix
```{r}
pred <- make.predictorMatrix(NSRI_df)
```

Then, we set the relevant columns in the predictor matrix to `0`, indicating that the corresponding column will not serve as a predictor in imputing the incomplete columns
```{r}
pred[, c("ID", "FinishDate", "IC1", "IC2", "D1", "V", "Group", "RRtest1", "RRtest2")] <- 0
```

We also set the rows for `IC2`, `RRtest1` and `RRtest2` to `0`, as these columns need not be imputed. 
```{r}
pred[c("IC2", "RRtest1", "RRtest2"), ] <- 0
```

---

## Imputation method
We also generate a default imputation method vector
```{r}
meth <- make.method(NSRI_df)
```
And set - for easthetic reasons - the methods for `IC2`, `RRtest1` and `RRtest2` to `""` (empty). This means that these columns have not been assigned an imputation method and are therefore excluded from imputation. 
```{r}
meth[c("IC2", "RRtest1", "RRtest2")] <- ""
meth
```

---

## Running the imputations
```{r cache=TRUE}
imp <- mice(NSRI_df, 
            predictorMatrix = pred, 
            method = meth, 
            m = 50, 
            maxit = 15, 
            print = FALSE)
```

---

# Exporting the workspace
We export the multiply imputed data set `imp` as an `R` workspace to the `\Workspaces\` directory in the project root. The resulting workspace name mirrors the name of this document. 
```{r}
save(imp, file = "Workspaces/3. Data imputation.Rdata")
```



---

End of document

---

```{r}
sessionInfo()
```
