==> documents/mt4/NSRI-main 3/4. Diagnostic evaluation.Rmd <==
---
title: "4. Diagnostic evaluation"
author: "Gerko Vink"
date: ""
bibliography: references.bib
output: 
   html_document:
    toc: true
    toc_depth: 5
    toc_float: true
    number_sections: false
---
  
<style type="text/css">
  
body{ /* Normal  */
  font-size: 12px;
  }
td {  /* Table  */
  font-size: 12px;
}
h1.title {
  font-size: 18px;
  color: DarkBlue;
}
h1 { /* Header 1 */
  font-size: 18px;
}
h2 { /* Header 2 */
  font-size: 18px;
}
h3 { /* Header 3 */
  font-size: 18px;
}
code.r{ /* Code block */
  font-size: 12px;
}
pre { /* Code block - determines code spacing between lines */
  font-size: 14px;
}
</style>

---

This document is aimed at evaluating the imputation procedure. 

---

# Packages used
For this document we use the following packages
```{r message=FALSE, warning=FALSE}
library(dplyr)    # Data manipulation
library(mice)     # Data imputation
library(magrittr) # Pipes
library(purrr)    # Functional programming
library(haven)    # Import data
library(DT)       # Interactive tables
```

---

# Read in the data file
We read in the workspace `3. Data imputation.Rdata` from the `\Workspaces\` directory in the project's root. This workspace has been generated by the document `3. Data imputation.Rmd`. The proces leading to this workspace is also documented in the file `3. Data imputation.html`.
```{r}
load("Workspaces/3. Data imputation.Rdata")
```

---

# Algorithmic convergence
To inspect the convergence of the mice algorithm. 
```{r}
plot(imp)
```

---

# Medians

## Evaluation function
```{r}
evaluate.median <- function(object, column, print = TRUE){
  observed <- object$data[, column] %>%  median(na.rm = TRUE)
  imputed <- complete(object, "all")  %>% 
    map(~ median(.[, column])) %>% 
    unlist
  if (print){
    plot(1:object$m, imputed, col = mdc(2), lwd = 4, main = column, ylab = "median", xlab = "imputation")
    abline(h =  observed, col = mdc(1), lwd = 2)
  }
  return(list(observed = observed, imputed = imputed))
}
```

## Medians per column
```{r}
which.loop <- names(imp$data)[11:85]
for (i in 1:length(which.loop)){
  evaluate.median(imp, which.loop[i])
}
```

---

# Tables

## Evaluation function
```{r}
evaluate.tables <- function(object, column, print = TRUE){
  observed <- table(imp$data[, column]) %>% prop.table() %>% round(digits = 4)
  imputed <- complete(object, "all")  %>% 
    map(~ table(.[, column]) %>% prop.table) %>% 
    Reduce("+", .)/object$m 
  print(list(column = column,
              observed = observed, 
              imputed = imputed %>% round(digits = 4), 
              bias = (observed - imputed) %>% round(digits = 4)))
}
```

## tables per column
```{r}
which.loop <- names(imp$data)[11:85]
for (i in 1:length(which.loop)){
  evaluate.tables(imp, which.loop[i])
}
```

# Densities
```{r}
densityplot(imp, ~ F1 + F2 + F3)
densityplot(imp, ~ F4 + F5 + F6)
densityplot(imp, ~ F7 + F8 + F9)
densityplot(imp, ~ F10 + F11 + F12)
densityplot(imp, ~ F13 + F14 + F15)
densityplot(imp, ~ F16 + F17 + F18)
densityplot(imp, ~ F19 + F20 + F21)
densityplot(imp, ~ F22 + F23 + F24)
densityplot(imp, ~ F25 + F26 + F27)
densityplot(imp, ~ F28 + F29 + F30)
densityplot(imp, ~ F31 + F32 + F33)
densityplot(imp, ~ F34 + F35 + F36)
densityplot(imp, ~ F37 + F38 + F39)
densityplot(imp, ~ F40 + F41 + F42)
densityplot(imp, ~ F43 + F44 + F45)
densityplot(imp, ~ F46 + F47 + F48)
densityplot(imp, ~ F49 + F50 + F51)
densityplot(imp, ~ F52 + F53 + F54)
densityplot(imp, ~ F55 + F56 + F57)
densityplot(imp, ~ F58 + F59 + F60)
densityplot(imp, ~ F61 + F62 + F63)
densityplot(imp, ~ F64 + F65 + F66)
densityplot(imp, ~ F67 + F68 + F69)
densityplot(imp, ~ F70 + F71 + F72)
densityplot(imp, ~ F73 + F74 + F75)
```

---

End of document

---

```{r}
sessionInfo()
```
